Syncing Up

+++ src/common/bilateralcl.h
@@ -35,7 +35,7 @@ typedef struct dt_bilateral_cl_t
 {
   dt_bilateral_cl_global_t *global;
   int devid;
-  size_t size_x, size_y, size_z;
+  int size_x, size_y, size_z;
   int width, height;
   size_t blocksizex, blocksizey;
   float sigma_s, sigma_r;
@@ -65,6 +65,9 @@ void dt_bilateral_free_cl_global(dt_bilateral_cl_global_t *b);
 
 #endif // HAVE_OPENCL
 
-// modelines: These editor modelines have been set for all relevant files by tools/update_modelines.sh
+// clang-format off
+// modelines: These editor modelines have been set for all relevant files by tools/update_modelines.py
 // vim: shiftwidth=2 expandtab tabstop=2 cindent
 // kate: tab-indents: off; indent-width 2; replace-tabs on; indent-mode cstyle; remove-trailing-spaces modified;
+// clang-format on
+
